+++ packages/plugins/Attributions/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unpublished
 
-- Feature: Add new configuration parameter `openLeft` to be able to configure the direction the attributions content opens in `renderType` `independent`.
+- Feature: Use the configured `layoutTag` to identify the direction the attributions content should open in `renderType` `independent`.
 
 ## 1.4.0